handle channel subcommand help from fluvio-channel

+++ crates/fluvio-channel-cli/src/cli/delete.rs
@@ -1,4 +1,4 @@
-use color_eyre::Result;
+use color_eyre::{Result, eyre::eyre};
 use fluvio_channel::FluvioChannelConfig;
 use std::path::PathBuf;
 use structopt::StructOpt;
@@ -8,14 +8,21 @@ use tracing::debug;
 pub struct DeleteOpt {
     #[structopt(long)]
     config: Option<PathBuf>,
-    channel: String,
+    channel: Option<String>,
     // binary-path
     // extension-path
     // image_tag_strategy
+    #[structopt(short, long)]
+    help: bool,
 }
 
 impl DeleteOpt {
     pub async fn process(&self) -> Result<()> {
+        if self.help {
+            let _ = DeleteOpt::clap().print_help();
+            println!();
+            return Ok(())
+        }
         // Open config file
 
         // Load in the config file
@@ -23,41 +30,49 @@ impl DeleteOpt {
 
         debug!("Looking for channel config");
 
-        let cli_config_path = if let Some(path) = &self.config {
-            debug!("Using provided channel config path");
-            path.to_path_buf()
-        } else {
-            debug!("Using default channel config path");
-            FluvioChannelConfig::default_config_location()
-        };
-
-        // Open file
-
-        if let Ok(mut load_config) = FluvioChannelConfig::from_file(cli_config_path) {
-            debug!("Loaded channel config");
-            let _ = if let Ok(Some(channel_info)) = load_config.remove_channel(self.channel.clone())
-            {
-                load_config.save()?;
-
-                debug!(
-                    "Delete binary and extensions dir for channel \"{}\"",
-                    self.channel
-                );
-
-                debug!("Deleting: {}", channel_info.get_binary_path().display());
-                std::fs::remove_file(channel_info.get_binary_path())?;
-
-                debug!("Deleting: {}", channel_info.get_extensions_path().display());
-                std::fs::remove_dir_all(channel_info.get_extensions_path())?;
-
-                println!("Deleted release channel \"{}\"", self.channel);
+        if let Some(channel_name) = &self.channel {
+            let cli_config_path = if let Some(path) = &self.config {
+                debug!("Using provided channel config path");
+                path.to_path_buf()
             } else {
-                println!("Release channel \"{}\" not found", self.channel);
+                debug!("Using default channel config path");
+                FluvioChannelConfig::default_config_location()
             };
-            Ok(())
+
+            // Open file
+
+            if let Ok(mut load_config) = FluvioChannelConfig::from_file(cli_config_path) {
+                debug!("Loaded channel config");
+                let _ = if let Ok(Some(channel_info)) =
+                    load_config.remove_channel(channel_name.clone())
+                {
+                    load_config.save()?;
+
+                    debug!(
+                        "Delete binary and extensions dir for channel \"{}\"",
+                        channel_name
+                    );
+
+                    debug!("Deleting: {}", channel_info.get_binary_path().display());
+                    std::fs::remove_file(channel_info.get_binary_path())?;
+
+                    debug!("Deleting: {}", channel_info.get_extensions_path().display());
+                    std::fs::remove_dir_all(channel_info.get_extensions_path())?;
+
+                    println!("Deleted release channel \"{}\"", channel_name);
+                } else {
+                    println!("Release channel \"{}\" not found", channel_name);
+                };
+                Ok(())
+            } else {
+                println!("No channel config found");
+                Ok(())
+            }
         } else {
-            println!("No channel config found");
-            Ok(())
+            println!("No channel name provided");
+            let _ = DeleteOpt::clap().print_help();
+            println!();
+            Err(eyre!(""))
         }
     }
 }
